Fix compatibility with FastAPI>0.90.1 and Pydantic >2.0
